requirements.txt
requests
pillow>=10.3.0
python-dotenv